+++ src/firebase/firebaseEnums.ts
@@ -1,0 +1,9 @@
+/**
+ * Firestore collection names
+ */
+export enum Collection {
+  USERS = "users",
+  COMMENTS = "comments",
+  QUESTIONS = "questions",
+  GRAPHS = "graphs",
+}
